requirements.txt
certifi==2019.3.9
chardet==3.0.4
distro==1.4.0
Django==2.1.7
get==2019.3.22
idna==2.8
numpy==1.16.2
pandas==0.24.2
pika==0.13.0
post==2019.3.22
public==2019.3.22
python-dateutil==2.8.0
pytz==2018.9
query-string==2019.3.22
request==2018.11.20
requests==2.21.0
six==1.12.0
tabula-py==1.3.1
tika==1.19
urllib3==1.24.1
